projects/02-nosql-agent/requirements.txt
pytest==7.4.4
python-dotenv==1.0.1
openai==2.0.1